# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "rust-bc REST API",
    "version": "1.0.0",
    "description": "REST API for rust-bc. **Gateway** routes wrap payloads in a common envelope (`status`, `status_code`, `message`, `data`, `error`, `timestamp`, `trace_id`). **Legacy** routes (wallets, contracts, billing, mining, etc.) still use `{ success, data?, message? }` until migrated."
  },
  "servers": [{ "url": "/api/v1", "description": "API v1" }],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional billing / rate-limit API key"
      }
    },
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "field": { "type": "string", "nullable": true }
        },
        "required": ["code", "message"]
      },
      "GatewayEnvelope": {
        "type": "object",
        "description": "Successful gateway response",
        "properties": {
          "status": { "type": "string", "enum": ["Success"] },
          "status_code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "OK" },
          "data": {},
          "error": { "type": "null" },
          "timestamp": { "type": "string", "format": "date-time" },
          "trace_id": { "type": "string", "format": "uuid" }
        },
        "required": ["status", "status_code", "message", "timestamp", "trace_id"]
      },
      "GatewayErrorEnvelope": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["Failure"] },
          "status_code": { "type": "integer" },
          "message": { "type": "string" },
          "data": { "type": "null" },
          "error": { "$ref": "#/components/schemas/ErrorDto" },
          "timestamp": { "type": "string", "format": "date-time" },
          "trace_id": { "type": "string" }
        }
      },
...
      "Block": {
        "type": "object",
        "additionalProperties": true,
        "description": "Block record (index, hash, transactions, merkle_root, ...)"
      },
      "CreateTransactionRequest": {
        "type": "object",
        "required": ["from", "to", "amount"],
        "properties": {
          "from": { "type": "string" },
          "to": { "type": "string" },
          "amount": { "type": "integer", "format": "uint64" },
          "fee": { "type": "integer", "format": "uint64" },
          "data": { "type": "string" },
          "signature": { "type": "string" }
        }
      },
      "CreateBlockRequest": {
        "type": "object",
        "required": ["transactions"],
        "properties": {
          "transactions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CreateTransactionRequest" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/blocks": {
      "get": {
        "tags": ["Gateway"],
        "summary": "List full chain",
        "responses": {
          "200": {
            "description": "Gateway envelope; `data` is Block[]",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/GatewayEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Block" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Gateway"],
        "summary": "Create / mine block from transaction list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBlockRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Gateway envelope; `data` is new block hash (string)",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/GatewayEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "type": "string", "description": "Block hash" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GatewayErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/blocks/index/{index}": {
      "get": {
        "tags": ["Gateway"],
        "summary": "Block by height",
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "uint64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Gateway envelope; `data` is Block",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/GatewayEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Block" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GatewayErrorEnvelope" }
              }
            }
          }
        }
      }
    },
    "/blocks/{hash}": {
      "get": {
        "tags": ["Gateway"],
        "summary": "Block by hash",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Gateway envelope; `data` is Block",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/GatewayEnvelope" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Block" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GatewayErrorEnvelope" }
              }
            }
          }
        }
      }
    },
...
  }
}